Removed `Handler` as a class with a static method and simply made a utils in `gcp` folder.
